Move PointStampedDisplay and delete its visual
- The visual doesn't do anything but delegate to Shape
- It was meant to be expanded, but that didn't happen in four years
