+++ rviz_default_plugins/src/rviz_default_plugins/displays/point/point_display.cpp
@@ -10,7 +10,7 @@
 
 #include "point_visual.h"
 
-#include "point_display.h"
+#include "point_display.hpp"
 
 namespace rviz
 {
